split particles from python
constexpr  nbRefinedPart on particle dispatcher from aggregate of patterns

Pharesee Particles split

Pharesee particles split returns new instance of self - pybind pyarray assertions in debug mode

finalize python splitting

what

correct ConstArray

drop deprecated StaticSamraiLifeCycle class, &&||| -> and|or
