+++ src/initializer/data_provider.h
@@ -40,7 +40,6 @@ namespace initializer
 
 
 
-    // template<std::size_t dim>
     using PHAREDict
         = cppdict::Dict<int, double, std::vector<double>, std::size_t, std::optional<std::size_t>,
                         std::string, ScalarFunction<1>, ScalarFunction<2>, ScalarFunction<3>>;
