+++ tests/simulator/__init__.py
@@ -157,15 +157,3 @@ def populate_simulation(dim, interp, **input):
 
     return simulation
 
-
-def cpp_splitter_type(dim, interp, n_particles):
-    from pybindlibs import cpp # includes static samrai lifecycle
-
-    return getattr(cpp,
-        "Splitter_" + str(dim) + "_" + str(interp)+ "_" + str(n_particles),
-    )
-
-
-def create_splitter(dim, interp, n_particles):
-
-    return cpp_splitter_type(dim, interp, n_particles)()
