+++ tests/simulator/diagnostics.py
@@ -52,12 +52,23 @@ class DiagnosticsTest(unittest.TestCase):
             self.simulator.initialize()
             self.simulator.diagnostics().dump(timestamp=0, timestep=1)
 
-            if dim == 1: # TODO REMOVE IF WHEN PHARESEE SUPPORTS 2D!
-                for diagInfo in ph.globals.sim.diagnostics:
-                    # diagInfo.quantity starts with a / this interferes with os.path.join, hence   [1:]
-                    h5_file = os.path.join(local_out, (diagInfo.quantity + ".h5").replace('/', '_')[1:])
-                    self.assertTrue(os.path.exists(h5_file))
-                    print("hier", hierarchy_from(h5_filename=h5_file))
+            for diagInfo in ph.globals.sim.diagnostics:
+                # diagInfo.quantity starts with a / this interferes with os.path.join, hence   [1:]
+                h5_file = os.path.join(local_out, (diagInfo.quantity + ".h5").replace('/', '_')[1:])
+                self.assertTrue(os.path.exists(h5_file))
+
+                print("h5_file", h5_file)
+                hier = hierarchy_from(h5_filename=h5_file)
+                if h5_file.endswith("domain.h5"):
+                    for patch in hier.level(0).patches:
+                        for qty_name, pd in patch.patch_datas.items():
+                            splits = pd.dataset.split(ph.globals.sim)
+                            self.assertTrue(splits.size() == pd.dataset.size() * 2)
+                            print("splits.iCell", splits.iCells)
+                            print("splits.delta", splits.deltas)
+                            print("splits.weight", splits.weights)
+                            print("splits.charge", splits.charges)
+                            print("splits.v", splits.v)
 
             self.simulator = None
